requirements.txt
customtkinter==5.2.2
Pillow==11.0.0
cryptography>=42.0.0
pyinstaller
openai>=1.40.0
google-generativeai
anthropic
groq
mistralai
openpyxl
requests
beautifulsoup4
playwright
playwright-stealth
huggingface_hub
pypdf
pandas
matplotlib
python-pptx
scrapegraphai
langchain
yfinance
diffusers
torch>=2.0.0
nest_asyncio
python-docx==1.2.0
pydantic==2.11.0
scrapegraph-py==1.44.0
chromadb==0.5.23
sentence-transformers==5.2.0
langchain-text-splitters==0.3.4